refactor(cleanup): extract adapter config parsers to adapter-utils (pass 8)
Moved parseCommaArgs, parseEnvVars, parseEnvBindings, and parseJsonObject
from 7 adapter ui/build-config.ts files into a new
packages/adapter-utils/src/config-parser.ts module, re-exported from the
package index.

Adapters updated: claude-local, codex-local, cursor-local, gemini-local,
opencode-local, pi-local, openclaw-gateway. Each now imports only the
parsers it actually uses from @paperclipai/adapter-utils.

Removes ~321 lines of byte-for-byte duplication (per .cleanup-reports/
01-dedup.md HIGH #3). Typecheck: all 20 workspace packages green.
