+++ packages/adapters/openclaw-gateway/src/ui/build-config.ts
@@ -1,16 +1,5 @@
 import type { CreateConfigValues } from "@paperclipai/adapter-utils";
-
-function parseJsonObject(text: string): Record<string, unknown> | null {
-  const trimmed = text.trim();
-  if (!trimmed) return null;
-  try {
-    const parsed = JSON.parse(trimmed);
-    if (typeof parsed !== "object" || parsed === null || Array.isArray(parsed)) return null;
-    return parsed as Record<string, unknown>;
-  } catch {
-    return null;
-  }
-}
+import { parseJsonObject } from "@paperclipai/adapter-utils";
 
 export function buildOpenClawGatewayConfig(v: CreateConfigValues): Record<string, unknown> {
   const ac: Record<string, unknown> = {};
